% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_pca.R
\name{multilevel_pca}
\alias{multilevel_pca}
\title{Multilelvel Principal Component Analysis}
\usage{
multilevel_pca(Y = NULL, id = NULL, twoway = TRUE,
  cov.method = c("m1", "m2"), pve = 0.9, tlength = ncol(Y),
  smoothing = FALSE, smooth.method = c("sf", "sc"), nk = 15)
}
\arguments{
\item{Y}{\code{matrix} with columns as variables/grid points, and rows as subject-visit, must be supplied.}

\item{id}{Must be supplied, a vector containing the id information used to identify clusters.}

\item{twoway}{logical, indicating whether to carry out twoway ANOVA and calculate
visit-specific means. Defaults to \code{FALSE}.}

\item{cov.method}{covariance estimation approaches.}

\item{pve}{proportion of variance explained: used to choose the number of principal components.}

\item{tlength}{For functional data,the length of the interval that the functions are
evaluated at. Defualt to be number of grid points which is exactly number of variables
for multivariate data.}

\item{smoothing}{logical, indicating whether smoothing should be carried out or not, only use this
for functional data.}

\item{smooth.method}{Method of smoothing, "sf" indicates directly smoothing the functions using gam.
"sc" indicates smoothing the covariance using penalized smoothing splines.}

\item{nk}{number of knots for the smoothing}
}
\value{
An object of class \code{mpca} containing:
\item{Y.df}{The original data.}
\item{mu}{estimated mean function.}
\item{eta}{the estimated visit specific shifts from overall mean.}
\item{npc}{number of PCs.}
\item{pctvar}{percentage of variation explained by the kept pcs within each level.}
\item{varofTot}{percentage of variation of the total variability explained by level 1 and level 2.}
\item{evectors}{level 1/2 eigen vectors/functions.}
\item{evalues}{level 1/2 eigen values.}
\item{scores}{level 1/2 pc scores.}
}
\description{
Conduct multilevel (functional) principal component analysis.
A mixed model framework is used to estimate scores and obtain variance estimates.
}
\examples{
\dontrun{
mpca.y = mfpca.sc(Y = Y, id = id, twoway = T, cov.method = "m1")
}
}
\references{
Di, C., Crainiceanu, C., Caffo, B., and Punjabi, N. (2009).
Multilevel functional principal component analysis. \emph{Annals of Applied
Statistics}, 3, 458--488.

Di, C., Crainiceanu, C., Caffo, B., and Punjabi, N. (2014).
Multilevel sparse functional principal component analysis. \emph{Stat}, 3, 126--143.

Goldsmith, J., Greven, S., and Crainiceanu, C. (2013). Corrected confidence
bands for functional data using principal components. \emph{Biometrics},
69(1), 41--51.
}
